+++ CHANGELOG.md
@@ -7,14 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-12-01
+
 ### Added
 
+- Code documentation
 - Added a reset preferences operator
 - Previous selection is now remembered and used when no selection
 
 ### Fixed
 
 - Selection is now taking into account all outliners and viewports
+- Hided Items are now unselected in order to fix some issues with a selection of hided and visible objects in outliners
 - Changed default hide method
 
 ## [1.2.3] - 2024-11-28